Fix warnings for 32-bit builds

+++ include/uxs/io/iflatbuf_impl.h
@@ -10,15 +10,15 @@ typename basic_iflatbuf<CharT>::pos_type basic_iflatbuf<CharT>::seekimpl(off_typ
     size_type pos = this->curr() - this->first();
     switch (dir) {
         case seekdir::beg: {
-            pos = off > 0 ? (static_cast<size_type>(off) < sz ? off : sz) : 0;
+            pos = static_cast<size_t>(off > 0 ? (static_cast<size_type>(off) < sz ? off : sz) : 0);
         } break;
         case seekdir::curr: {
             if (off == 0) { return pos; }
-            pos = off > 0 ? (static_cast<size_type>(off) < sz - pos ? pos + off : sz) :
-                            (static_cast<size_type>(-off) < pos ? pos + off : 0);
+            pos = static_cast<size_t>(off > 0 ? (static_cast<size_type>(off) < sz - pos ? pos + off : sz) :
+                                                (static_cast<size_type>(-off) < pos ? pos + off : 0));
         } break;
         case seekdir::end: {
-            pos = off < 0 ? (static_cast<size_type>(-off) < sz ? sz + off : 0) : sz;
+            pos = static_cast<size_t>(off < 0 ? (static_cast<size_type>(-off) < sz ? sz + off : 0) : sz);
         } break;
     }
     this->setcurr(this->first() + pos);
